메인 통합 - 세션값에 따라 채용공고or이력서 리스트 보여주기
