--- src/main/webapp/WEB-INF/views/company/main.jsp
@@ -1,91 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-	pageEncoding="UTF-8"%>
-<%@ include file="../layout/header.jsp"%>
-
-<div class="container mt-3" style="border: 3px solid #000000">
-	<div class="row row-cols-1 row-cols-md-3 g-4">
-		<div id="box1" class="col">
-			<div class="card h-100">
-				<img src="/image/1.jpg" class="card-img-top" alt="...">
-				<div class="card-body">
-					<h5 class="card-title">Card title</h5>
-				</div>
-			</div>
-		</div>
-		<div id="box2" class="col">
-			<div class="card h-100">
-				<img src="/image/1.jpg" class="card-img-top" alt="...">
-				<div class="card-body">
-					<h5 class="card-title">Card title</h5>
-				</div>
-			</div>
-		</div>
-		<div id="box3" class="col">
-			<div class="card h-100">
-				<img src="/image/1.jpg" class="card-img-top" alt="...">
-				<div class="card-body">
-					<h5 class="card-title">Card title</h5>
-				</div>
-			</div>
-		</div>
-	</div>
-</div>
-<div class="mb-3"></div>
-<div class="dropdown">
-</div>
-
-
-<div class = "d-flex justify-content-end" >
-	<div style="width:300px">
-		<form class="d-flex" method="get" action="/companymain">
-			<input class="form-control me-2" type="text" placeholder="Search" name="keyword">
-			<button class="btn btn-primary" type="submit">Search</button>
-		</form>
-	</div>
-</div>
-
-<div class="d-flex justify-content-center">
-	<table class="table table-bordered" style="text-align: center">
-		<thead>
-			<tr>
-				<th>번호</th>
-				<th>이력서 제목</th>
-				<th>희망근무지역</th>
-			</tr>
-		</thead>	
-		<tbody id="table">
-		<c:forEach var="resumesList" items="${resumesList}">
-			<tr>
-				<td>${resumesList.resumesId}</td>
-				<td>	
-					<div id="apply" class="container p-4 my-4 border">
-					<a href="/personal/resumes/${resumesList.resumesId}">
-					${resumesList.resumesTitle}</a>
-					</div>
-				</td>
-				<td>${resumesList.resumesPlace}</td>
-			</tr>
-		</c:forEach>
-		</tbody>
-	</table>
-</div>
-
-	<div class="d-flex justify-content-center">
-		<ul class="pagination">
-		   <li class='page-item ${paging.first ? "disabled" : " "}'>
-		   		<a class="page-link" href="?page=${paging.currentPage-1}&keyword=${paging.keyword}">Prev</a>
-		   </li>
-		      
-		   <c:forEach var="num" begin="${paging.startPageNum}" end="${paging.lastPageNum}" step="1">
-		      <li class='page-item ${paging.currentPage == num-1 ? "active" : ""}'>
-		      		<a class="page-link" href="?page=${num-1}&keyword=${paging.keyword}">${num}</a>
-		      </li>
-		   </c:forEach>
-		   
-		   <li class='page-item ${paging.last ? "disabled" : " "}'>
-		   		<a class="page-link" href="?page=${paging.currentPage+1}&keyword=${paging.keyword}">Next</a>
-		   </li>
-		</ul>	
-	</div>
-
-<%@ include file="../layout/footer.jsp"%>
