+++ src/main/webapp/WEB-INF/views/layout/header.jsp
@@ -58,7 +58,6 @@
                            <li class="nav-item"><a class="nav-link" href="/loginForm">로그인/회원가입</a></li>
                         </c:when>
                         <c:when test="${principal.companyId != null}">
-                           <li class="nav-item"><a class="nav-link" href="/companymain">기업 메인화면</a></li>
                            <li class="nav-item"><a class="nav-link" href="/recommend">관심 이력서 보기</a></li>
                            <li class="nav-item"><a class="nav-link" href="/company/jobPostingBoardList"> 공고관리</a></li>
                            <li class="nav-item"><a class="nav-link" href="/company/inform">기업정보</a></li>
